requirements.txt
Flask==3.0.3
flask-cors==4.0.1
binance-connector==3.12.0
python-dotenv==1.0.1
pandas==2.2.1
pymongo==4.6.1
flask-login==0.6.3
werkzeug==3.0.3
bcrypt==4.1.2
